refactor: mv toSentenceCase to rule module
